Changed #include "lpc17xx.h" to #include "LPC17xx.h" in lpc17xx_clkpwr.h Changed #include "lpc17xx.h" to #include "LPC17xx.h" in lpc17xx_pinsel.h Undefine macros for TRUE and FALSE in lpc_types.h before the Bool typedef. Changed float to uint32_t in lpc17xx_systick.c. Defined CHECK_PARAM() as assert() in lpc17xx_libcfg_default.h.
git-svn-id: svn+ssh://subversion.munts.net/src/trunk/ARM@3514 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ lpc17xx/libs/lpc17xx.cmsis.driver.library/Drivers/source/lpc17xx_systick.c
@@ -58,7 +58,7 @@
 void SYSTICK_InternalInit(uint32_t time)
 {
 	uint32_t cclk;
-	float maxtime;
+	uint32_t maxtime;
 
 	cclk = SystemCoreClock;
 	/* With internal CPU clock frequency for LPC17xx is 'SystemCoreClock'
@@ -91,7 +91,7 @@ void SYSTICK_InternalInit(uint32_t time)
  **********************************************************************/
 void SYSTICK_ExternalInit(uint32_t freq, uint32_t time)
 {
-	float maxtime;
+	uint32_t maxtime;
 
 	/* With external clock frequency for LPC17xx is 'freq'
 	 * And limit 24 bit for RELOAD value
